+++ index.js
@@ -1,4 +1,4 @@
 'use strict';
 
 // Export the handler constructor
-module.exports = require('./lib/handler');
+module.exports = require('./lib/aftership');
