+++ lib/error/error.js
@@ -32,18 +32,33 @@ class AftershipError extends Error {
 	 * @param retry_count {number}
 	 * @constructor
 	 */
-	static getSdkError(error_enum, error_data, retry_count) {
+	static getSdkError(error_enum, error_data) {
 		let payload = {
 			type: error_enum.type,
 			message: error_enum.message,
 			data: error_data
 		};
 
+		return new AftershipError('SDK', payload);
+	}
+
+	/**
+	 * Static Method for getting SDK error
+	 *
+	 * @param error_enum {object}
+	 * @param error_data {object} - The object trigger the error
+	 * @param retry_count {number}
+	 * @constructor
+	 */
+	static getRequestError(error, error_data, retry_count) {
+		error.data = error_data;
+		error.type = error.code;
+
 		if (retry_count) {
-			payload.retry_count = retry_count;
+			error.retry_count = retry_count;
 		}
 
-		return new AftershipError('SDK', payload);
+		return error;
 	}
 
 	/**
