+++ lib/error/error_enum.js
@@ -5,16 +5,10 @@
  */
 const ErrorType = {
 	ConstructorError: 'ConstructorError',
-	HandlerError: 'HandlerError',
-	ETIMEDOUT: 'ETIMEDOUT'
+	HandlerError: 'HandlerError'
 };
 
 const ErrorEnum = {
-	// ConnectionTimeout
-	ConnectionTimeout: {
-		type: ErrorType.ETIMEDOUT,
-		message: 'Connection Timeout'
-	},
 	// ConstructorError
 	ConstructorInvalidApiKey: {
 		type: ErrorType.ConstructorError,
