+++ lib/payload.js
@@ -1,0 +1,101 @@
+'use strict';
+
+// Declare imports
+const _ = require('lodash');
+const AftershipError = require('./error/error');
+const ErrorEnum = require('./error/error_enum');
+
+// Global variable
+const methods = ['GET', 'POST', 'DELETE', 'PUT'];
+
+// Payload class
+class Payload {
+	/**
+     * Function to create the request_object and other information
+     *
+	 * @param aftership {Aftership} - the aftership object
+	 * @param method {string} - get, post, put or delete
+     * @param path {string} - pathname for URL
+	 * @param options {object}
+     *	body {object} - POST body
+     *	query {object} - query object
+	 *	retry {boolean} - Retry if fail? override this.retry if set
+	 *	raw {boolean} - if true, return string, else return object, default false
+	 * @return the payload object with properities below
+	 * 	{
+	 *		request_object {object} - object to create request
+	 *		retry {boolean} - retry flag
+	 *		raw {boolean} - response with string or not
+ 	 *	}
+     */
+	constructor(aftership, method, path, options) {
+		// body, query, retry, raw
+		if (methods.indexOf(method.toUpperCase()) === -1) {
+			// Verify method
+			throw AftershipError.getSdkError(ErrorEnum.HandlerInvalidMethod, method);
+		} else if (!_.isString(path)) {
+			// Verify path
+			throw AftershipError.getSdkError(ErrorEnum.HandlerInvalidPath, path);
+		} else if (!_.isNull(options) && !_.isUndefined(options) && !_.isPlainObject(options)) {
+			// Verify options
+			throw AftershipError.getSdkError(ErrorEnum.HandlerInvalidOptions, options);
+		}
+
+		// Verify options value
+		options = options || {};
+		if (!_.isNull(options.body) && !_.isUndefined(options.body) && !_.isPlainObject(options.body)) {
+			// Verify body
+			throw AftershipError.getSdkError(ErrorEnum.HandlerInvalidBody, options.body);
+		} else if (!_.isNull(options.query) && !_.isUndefined(options.query) && !_.isPlainObject(options.query)) {
+			// Verify query
+			throw AftershipError.getSdkError(ErrorEnum.HandlerInvalidQuery, options.query);
+		} else if (!_.isNull(options.retry) && !_.isUndefined(options.retry) && !_.isBoolean(options.retry)) {
+			// Verify retry
+			throw AftershipError.getSdkError(ErrorEnum.HandlerInvalidRetry, options.retry);
+		} else if (!_.isNull(options.raw) && !_.isUndefined(options.raw) && !_.isBoolean(options.raw)) {
+			// Verify raw
+			throw AftershipError.getSdkError(ErrorEnum.HandlerInvalidRaw, options.raw);
+		}
+
+		// Request object
+		let request_object = {
+			headers: {
+				'aftership-api-key': aftership.api_key,
+				'Content-Type': 'application/json',
+				'x-aftership-agent': 'LANGUAGE-sdk-VERSION'
+			},
+			url: aftership.endpoint + path,
+			method: method,
+			body: options.body,
+			qs: options.query,
+			proxy: aftership.proxy,
+			json: true
+		};
+
+		// Remove all null or undefined
+		for (let key in request_object) {
+			if (request_object[key] === null || request_object[key] === undefined) {
+				delete request_object[key];
+			}
+		}
+
+		// Payload
+		let payload = {
+			request_object: request_object,
+			retry: (!_.isNull(options.retry) && !_.isUndefined(options.retry)) ? options.retry : aftership.retry,
+			raw: (!_.isNull(options.raw) && !_.isUndefined(options.raw)) ? options.raw : false
+		};
+
+		// Add retry count if needed
+		if (payload.retry) {
+			payload.retry_count = 0;
+		}
+
+		return payload;
+	}
+}
+
+// Exports the constructor
+module.exports = function (api_key, endpoint, proxy, retry) {
+	return new Payload(api_key, endpoint, proxy, retry);
+};
